+++ src/store/modules/normalizers.js
@@ -71,18 +71,16 @@ export const parseTuples = (tuples, key) => {
         return [...acc, { [mascot.tuple[0]]: { ...mascot.tuple[1], id: `f${(~~(Math.random() * 1e8)).toString(16)}` }}]
       }, [])
     } else if (Array.isArray(item.tuple[1]) &&
-      (item.tuple[0] === ':groups' || item.tuple[0] === ':replace' || item.tuple[0] === ':retries')) {
-      accum[item.tuple[0]] = item.tuple[1].reduce((acc, group) => {
-        return [...acc, { [group.tuple[0]]: { value: group.tuple[1], id: `f${(~~(Math.random() * 1e8)).toString(16)}` }}]
-      }, [])
-    } else if (item.tuple[0] === ':crontab') {
-      accum[item.tuple[0]] = item.tuple[1].reduce((acc, group) => {
-        return { ...acc, [group.tuple[1]]: group.tuple[0] }
-      }, {})
-    } else if (item.tuple[0] === ':match_actor') {
-      accum[item.tuple[0]] = Object.keys(item.tuple[1]).reduce((acc, regex) => {
-        return [...acc, { [regex]: { value: item.tuple[1][regex], id: `f${(~~(Math.random() * 1e8)).toString(16)}` }}]
-      }, [])
+      (item.tuple[0] === ':groups' || item.tuple[0] === ':replace' || item.tuple[0] === ':retries' || item.tuple[0] === ':headers' || item.tuple[0] === ':crontab')) {
+      if (item.tuple[0] === ':crontab') {
+        accum[item.tuple[0]] = item.tuple[1].reduce((acc, group) => {
+          return [...acc, { [group.tuple[1]]: { value: group.tuple[0], id: `f${(~~(Math.random() * 1e8)).toString(16)}` }}]
+        }, [])
+      } else {
+        accum[item.tuple[0]] = item.tuple[1].reduce((acc, group) => {
+          return [...acc, { [group.tuple[0]]: { value: group.tuple[1], id: `f${(~~(Math.random() * 1e8)).toString(16)}` }}]
+        }, [])
+      }
     } else if (item.tuple[0] === ':icons') {
       accum[item.tuple[0]] = item.tuple[1].map(icon => {
         return Object.keys(icon).map(name => {
@@ -103,7 +101,13 @@ export const parseTuples = (tuples, key) => {
     } else if (item.tuple[0] === ':ip') {
       accum[item.tuple[0]] = item.tuple[1].tuple.join('.')
     } else if (item.tuple[1] && typeof item.tuple[1] === 'object') {
-      accum[item.tuple[0]] = parseObject(item.tuple[1])
+      if (item.tuple[0] === ':params' || item.tuple[0] === ':match_actor') {
+        accum[item.tuple[0]] = Object.keys(item.tuple[1]).reduce((acc, key) => {
+          return [...acc, { [key]: { value: item.tuple[1][key], id: `f${(~~(Math.random() * 1e8)).toString(16)}` }}]
+        }, [])
+      } else {
+        accum[item.tuple[0]] = parseObject(item.tuple[1])
+      }
     } else {
       accum[item.tuple[0]] = item.tuple[1]
     }
@@ -214,11 +218,11 @@ export const wrapUpdatedSettings = (group, settings, currentState) => {
 const wrapValues = (settings, currentState) => {
   return Object.keys(settings).map(setting => {
     const [type, value] = settings[setting]
-    if (
-      type === 'keyword' ||
-      type.includes('keyword') ||
-      type.includes('tuple') && type.includes('list') ||
-      setting === ':replace'
+    if (type === 'keyword' ||
+      (Array.isArray(type) && (
+        type.includes('keyword') ||
+        (type.includes('tuple') && type.includes('list'))
+      ))
     ) {
       return { 'tuple': [setting, wrapValues(value, currentState)] }
     } else if (prependWithСolon(type, value)) {
@@ -231,15 +235,16 @@ const wrapValues = (settings, currentState) => {
       return { 'tuple': [value, setting] }
     } else if (type === 'map') {
       const mapValue = Object.keys(value).reduce((acc, key) => {
-        acc[key] = setting === ':match_actor' ? value[key] : value[key][1]
+        acc[key] = value[key][1]
         return acc
       }, {})
-      const mapCurrentState = setting === ':match_actor'
-        ? currentState[setting].reduce((acc, element) => {
-          return { ...acc, ...{ [Object.keys(element)[0]]: Object.values(element)[0].value }}
-        }, {})
-        : currentState[setting]
-      return { 'tuple': [setting, { ...mapCurrentState, ...mapValue }] }
+      return { 'tuple': [setting, { ...currentState[setting], ...mapValue }] }
+    } else if (type.includes('map')) {
+      const mapValue = Object.keys(value).reduce((acc, key) => {
+        acc[key] = value[key][1]
+        return acc
+      }, {})
+      return { 'tuple': [setting, mapValue] }
     } else if (setting === ':ip') {
       const ip = value.split('.').map(s => parseInt(s, 10))
       return { 'tuple': [setting, { 'tuple': ip }] }