+++ src/views/settings/components/Setting.vue
@@ -122,7 +122,7 @@ export default {
       return type === 'keyword' ||
         type === 'map' ||
         type.includes('keyword') ||
-        key === ':replace'
+        type.includes('map')
     },
     divideSetting(key) {
       return [':sslopts', ':tlsopts', ':adapter', ':poll_limits', ':queues', ':styling', ':invalidation', ':multi_factor_authentication'].includes(key)