+++ src/views/settings/components/inputComponents/index.js
@@ -1,5 +1,4 @@
 export { default as EditableKeywordInput } from './EditableKeywordInput'
-export { default as CrontabInput } from './CrontabInput'
 export { default as IconsInput } from './IconsInput'
 export { default as ImageUploadInput } from './ImageUploadInput'
 export { default as LinkFormatterInput } from './LinkFormatterInput'