+++ src/views/settings/index.vue
@@ -200,7 +200,7 @@ export default {
       return this.$store.state.settings.searchData
     },
     tabs() {
-      return tabs
+      return tabs(this.$store.state.settings.description)
     }
   },
   mounted: function() {