+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support pagination of local emoji packs and files
 - Add MRF Activity Expiration setting
 - Add ability to disable multi-factor authentication for a user
+- Add ability to manually evict and ban URLs from the Pleroma MediaProxy cache
+- Add Invalidation settings on MediaProxy tab
 - Ability to configure S3 settings on Upload tab
 - Show number of open reports in Sidebar Menu
 - Add confirmation message when deleting a user
@@ -30,6 +32,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Remove ability to moderate users that don't have valid nickname
 - Displays both labels and description in the header of group of settiings
 - Ability to add custom values in Pleroma.Upload.Filter.Mogrify setting
+- Change types of the following settings: ':groups', ':replace', ':federated_timeline_removal', ':reject', ':match_actor'. Update functions that parses and wraps settings data according to this change.
+- Move rendering Crontab setting from a separate component to EditableKeyword component
 - Show only those MRF settings that have been enabled in MRF Policies setting
 - Move Auto Linker settings to Link Formatter Tab as its configuration was moved to :pleroma, Pleroma.Formatter